@@ -165,10 +168,10 @@
     "ts-jest": "26.5.6",
     "ts-node": "^10.9.1",
     "tslint": "6.1.3",
-    "typescript": "4.6",
+    "typescript": "4.9",
     "url-loader": "4.1.1",
     "webpack": "5.33.2",
-    "webpack-bundle-analyzer": "4.4.1",
+    "webpack-bundle-analyzer": "4.8.0",
     "webpack-cli": "4.10.0",
     "webpack-dev-server": "3.11.2",
     "webpack-merge": "5.7.3",
