@@ -1,44 +1,39 @@
-const fs = require('fs');
-const path = require('path');
-const crypto = require('crypto');
+import os from 'os';
+import path from 'path';
 
-const { Platform, build } = require('electron-builder');
-const { notarize } = require('electron-notarize');
+import { CliOptions, build } from 'electron-builder';
+import { notarize } from 'electron-notarize';
+
 
 async function notarizing(context) {
   const { electronPlatformName, appOutDir } = context;
   if (electronPlatformName !== 'darwin' || process.env.DONT_SIGN_APP) return;
+  if (!process.env.APPLEID || !process.env.APPLEIDPASS || !process.env.APPLETEAMID) {
+    console.error([
+      'Error: No env variables `APPLEID`, `APPLEIDPASS`, `APPLETEAMID` are set.',
+      'To skip signing set `DONT_SIGN_APP=1` explicitly.',
+    ].join('\n'));
+    process.exit(1);
+  }
 
   const appName = context.packager.appInfo.productFilename;
 
   return await notarize({
+    tool: 'notarytool',
     appBundleId: 'com.spacemesh.wallet',
     appPath: `${appOutDir}/${appName}.app`,
     appleId: process.env.APPLEID,
     appleIdPassword: process.env.APPLEIDPASS,
+    teamId: process.env.APPLETEAMID,
   });
 }
 
 const args = process.argv.slice(2);
 if (args.length < 2 || args[0] !== '--target' || !['mac', 'windows', 'linux', 'mwl'].includes(args[1])) {
-  throw new Error("No valid flags provided. Usage example: 'node ./scripts/packagerScript.js --target {mac|linux|windows|mwl}'");
+  console.error("\nError: No valid flags provided. \nUsage example: 'node ./scripts/packagerScript.js --target {mac|linux|windows|mwl}'");
+  process.exit(1);
 }
 const targets = args[1] === 'mwl' ? ['mac', 'windows', 'linux'] : [args[1]];
-
-const fileHashList = {
-  dmg: {
-    file: 'dmg_installer',
-    hash: 'dmg_sha512'
-  },
-  exe: {
-    file: 'exe_installer',
-    hash: 'exe_sha512'
-  },
-  deb: {
-    file: 'deb_installer',
-    hash: 'deb_sha512'
-  }
-};
 
 const nodeFiles = {
   mac: { from: path.resolve('node/mac/'), to: 'node/' },
@@ -46,47 +41,8 @@ const nodeFiles = {
   linux: { from: path.resolve('node/linux/'), to: 'node/' }
 };
 
-const artifactsToPublishFile = path.join(__dirname, '..', 'release', 'publishFilesList.json');
-try {
-  const dirname = path.dirname(artifactsToPublishFile);
-  if (!fs.existsSync(dirname)) {
-    fs.mkdirSync(dirname);
-  }
-  if (!fs.existsSync(artifactsToPublishFile)) {
-    fs.writeFileSync(artifactsToPublishFile, JSON.stringify(fileHashList));
-  }
-} catch (error) {
-  console.error(error);
-  process.exit(1);
-}
-
-const getFileHash = ({ filename }) => {
-  const shaSum = crypto.createHash('sha512');
-  const fileContent = fs.readFileSync(filename);
-  const hashSum = shaSum.update(fileContent);
-  return hashSum.digest('hex');
-};
-
-const compileHashListFile = ({ artifactsToPublishFile, artifactPaths }) => {
-  const acceptedSuffixes = ['dmg', 'exe', 'deb', 'snap', 'AppImage'];
-  let hashList = fs.readFileSync(artifactsToPublishFile, 'utf8');
-  for (const fullPath of artifactPaths) {
-    const artifactSuffix = fullPath.split('.').pop();
-    const artifactName = fullPath.split('/').pop();
-    // installers only
-    if (acceptedSuffixes.indexOf(artifactSuffix) >= 0) {
-      const hash = getFileHash({ filename: fullPath });
-      const artifactNameSpacesReplaced = artifactName.replace(/ /g, '+');
-      hashList = hashList.replace(`${artifactSuffix}_installer`, artifactNameSpacesReplaced);
-      hashList = hashList.replace(`${artifactSuffix}_sha512`, hash);
-    }
-  }
-  fs.writeFileSync(artifactsToPublishFile, hashList);
-};
-
 const getBuildOptions = ({ target }) => {
-  const buildOptions = {
-    targets: Platform[target.toUpperCase()].createTarget(),
+  const buildOptions: Partial<CliOptions> = {
     publish: 'never',
     config: {
       appId: 'com.spacemesh.wallet',
@@ -108,6 +64,7 @@ const getBuildOptions = ({ target }) => {
       ],
       extraResources: [
         'resources/icons/*',
+        'resources/icon.icns'
       ],
       extraFiles: [
         nodeFiles[target],
@@ -118,20 +75,24 @@ const getBuildOptions = ({ target }) => {
         gatekeeperAssess: false,
         entitlements: path.join(__dirname, 'entitlements.mac.plist'),
         entitlementsInherit: path.join(__dirname, 'entitlements.mac.plist'),
-        target: ['zip', 'dmg'],
+        target: {
+          target: 'default',
+          arch: [os.arch() === 'arm64' ? 'arm64' : 'x64'],
+        },
         icon: path.join(__dirname, '..', 'resources', 'icon.icns'),
         binaries: [
           path.join(__dirname, '../node/mac/go-spacemesh'),
           path.join(__dirname, '../node/mac/libgpu-setup.dylib'),
           path.join(__dirname, '../node/mac/libMoltenVK.dylib'),
           path.join(__dirname, '../node/mac/libvulkan.1.dylib')
-        ]
+        ],
+        ...(process.env.DONT_SIGN_APP ? { identity: null } : {}),
       },
       dmg: {
         sign: false,
         window: {
-          width: '400',
-          height: '380'
+          width: 400,
+          height: 380
         },
         background: path.join(__dirname, '..', 'resources', 'background.png'),
         icon: path.join(__dirname, '..', 'resources', 'icon.icns'),
@@ -151,6 +112,9 @@ const getBuildOptions = ({ target }) => {
         title: 'Spacemesh'
       },
       win: {
+        ...(process.env.DONT_SIGN_APP ? {} : {
+          publisherName: 'Spacemesh (Unruly Technologies, Inc.)',
+        }),
         target: 'nsis',
       },
       nsis: {
@@ -166,23 +130,24 @@ const getBuildOptions = ({ target }) => {
         uninstallDisplayName: 'Spacemesh (${version})'
       },
       linux: {
-        target: ['deb'],
+        target: [
+          {
+            target: 'deb',
+            arch: [os.arch() === 'arm64' ? 'arm64' : 'x64'],
+          },
+          {
+            target: 'AppImage',
+            arch: [os.arch() === 'arm64' ? 'arm64' : 'x64'],
+          }
+        ],
         category: 'Utility',
-        icon: path.join(__dirname, '..', 'resources', 'icons', '512x512.png')
+        icon: path.join(__dirname, '..', 'resources', 'icon.icns')
       },
       directories: {
         buildResources: path.join(__dirname, '..', 'resources'),
         output: path.join(__dirname, '..', 'release')
       },
-      afterAllArtifactBuild: (buildResult) => {
-        try {
-          compileHashListFile({ artifactsToPublishFile, artifactPaths: buildResult.artifactPaths });
-          return [artifactsToPublishFile];
-        } catch (error) {
-          console.error(error.message);
-          process.exit(1);
-        }
-      },
+      afterSign: target === 'mac' ? notarizing : null,
       npmRebuild: false,
       publish: {
         provider: 'generic',
@@ -191,11 +156,6 @@ const getBuildOptions = ({ target }) => {
       },
     }
   };
-
-  if (target === 'mac') {
-    buildOptions.config.afterSign = notarizing;
-  }
-
   return buildOptions;
 };
 
