@@ -1,5 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
+import { useSelector } from 'react-redux';
+import ReactPortal from '../../screens/modal/ReactPortal';
+import { RootState } from '../../types';
 import Modal from './Modal';
 
 const Spinner = styled.img.attrs(({ theme: { icons } }) => ({
@@ -9,10 +12,20 @@ const Spinner = styled.img.attrs(({ theme: { icons } }) => ({
   height: 40px;
 `;
 
-const CloseAppModal = () => (
-  <Modal header="SHUTTING DOWN" subHeader="please wait...">
-    <Spinner alt="Please wait..." />
-  </Modal>
-);
+const CloseAppModal = () => {
+  const isClosingApp = useSelector((state: RootState) => state.ui.isClosingApp);
+
+  if (!isClosingApp) {
+    return null;
+  }
+
+  return (
+    <ReactPortal modalId="spacemesh-shutting-down">
+      <Modal header="SHUTTING DOWN" subHeader="please wait...">
+        <Spinner alt="Please wait..." />
+      </Modal>
+    </ReactPortal>
+  );
+};
 
 export default CloseAppModal;
