@@ -1,5 +1,5 @@
 import { BrowserWindow } from 'electron';
-import { Subject, withLatestFrom } from 'rxjs';
+import { BehaviorSubject, Subject, withLatestFrom } from 'rxjs';
 import promptBeforeClose from '../promptBeforeClose';
 import Logger from '../../logger';
 import { makeSubscription } from '../rx.utils';
@@ -11,18 +11,17 @@ export default (
   $quit: Subject<Electron.IpcMainEvent>,
   $managers: Subject<Managers>,
   $mainWindow: Subject<BrowserWindow>,
-  $isAppClosing: Subject<boolean>,
+  $isAppClosing: BehaviorSubject<boolean>,
   $showWindowOnLoad: Subject<boolean>
 ) =>
   makeSubscription(
-    $quit.pipe(withLatestFrom($mainWindow), withLatestFrom($managers)),
-    ([[event, mw], managers]) => {
+    $quit.pipe(withLatestFrom($mainWindow, $managers)),
+    ([event, mw, managers]) =>
       mw &&
-        promptBeforeClose(
-          mw,
-          managers || {},
-          $isAppClosing,
-          $showWindowOnLoad
-        )(event).catch((err) => logger.error('promptBeforeClose', err));
-    }
+      promptBeforeClose(
+        mw,
+        managers || {},
+        $isAppClosing,
+        $showWindowOnLoad
+      )(event).catch((err) => logger.error('promptBeforeClose', err))
   );
