@@ -1,40 +1,18 @@
 import { generateGenesisID } from '../../desktop/main/Networks';
-import { toHexString } from '../../shared/utils';
 
 describe('test Network util functions', () => {
   const genesisTime = '2022-11-20T20:00:00.498Z';
   const extraData = 'unique-testnet-identifier';
 
-  const expectedGenesisId = Uint8Array.from([
-    51,
-    13,
-    17,
-    59,
-    152,
-    12,
-    243,
-    163,
-    29,
-    125,
-    73,
-    182,
-    3,
-    150,
-    80,
-    133,
-    46,
-    74,
-    42,
-    216,
-  ]);
+  const expectedGenesisId = 'b92736fb23dfe78efe03c2c45fb638b9fe3afa70';
 
   it('test for generateGenesisID', () => {
     expect(generateGenesisID(genesisTime, extraData)).toEqual(
       generateGenesisID(genesisTime, extraData)
     );
 
-    expect(toHexString(expectedGenesisId)).toEqual(
-      generateGenesisID(genesisTime, extraData)
+    expect(generateGenesisID(genesisTime, extraData)).toEqual(
+      expectedGenesisId
     );
   });
 });
